+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 2.5.35
+
+### Patch Changes
+
+- Updated dependencies [6d71400]
+    - @sap-ux/ui5-application-writer@1.7.4
+    - @sap-ux/odata-service-writer@0.29.10
+    - @sap-ux/ui5-test-writer@0.7.57
+    - @sap-ux/cap-config-writer@0.12.50
+    - @sap-ux/fiori-generator-shared@0.13.59
+
+## 2.5.34
+
+### Patch Changes
+
+- Updated dependencies [be6ea11]
+    - @sap-ux/ui5-application-writer@1.7.3
+    - @sap-ux/ui5-test-writer@0.7.56
+    - @sap-ux/fiori-generator-shared@0.13.58
+    - @sap-ux/ui5-config@0.29.13
+    - @sap-ux/cap-config-writer@0.12.49
+
+## 2.5.33
+
+### Patch Changes
+
+- @sap-ux/fiori-generator-shared@0.13.57
+- @sap-ux/cap-config-writer@0.12.48
+
+## 2.5.32
+
+### Patch Changes
+
+- d11943d: fix(deps): update dependency i18next to v25.8.0
+- Updated dependencies [d11943d]
+    - @sap-ux/cap-config-writer@0.12.47
+    - @sap-ux/fiori-generator-shared@0.13.56
+    - @sap-ux/odata-service-writer@0.29.9
+    - @sap-ux/ui5-application-writer@1.7.2
+
+## 2.5.31
+
+### Patch Changes
+
+- Updated dependencies [924e26f]
+    - @sap-ux/ui5-test-writer@0.7.55
+
+## 2.5.30
+
+### Patch Changes
+
+- @sap-ux/fiori-generator-shared@0.13.55
+- @sap-ux/ui5-config@0.29.13
+- @sap-ux/cap-config-writer@0.12.46
+
 ## 2.5.29
 
 ### Patch Changes