+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.9.42
+
+### Patch Changes
+
+- Updated dependencies [6d71400]
+    - @sap-ux/project-access@1.34.3
+
+## 0.9.41
+
+### Patch Changes
+
+- d11943d: fix(deps): update dependency i18next to v25.8.0
+
 ## 0.9.40
 
 ### Patch Changes