+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/odata-service-writer
 
+## 0.29.10
+
+### Patch Changes
+
+- 6d71400: Changes to support v4.01 odata services
+- Updated dependencies [6d71400]
+    - @sap-ux/project-access@1.34.3
+    - @sap-ux/mockserver-config-writer@0.9.42
+
+## 0.29.9
+
+### Patch Changes
+
+- d11943d: fix(deps): update dependency i18next to v25.8.0
+- Updated dependencies [d11943d]
+    - @sap-ux/mockserver-config-writer@0.9.41
+
 ## 0.29.8
 
 ### Patch Changes