+++ packages/project-input-validator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/project-input-validator
 
+## 0.6.47
+
+### Patch Changes
+
+- Updated dependencies [6d71400]
+    - @sap-ux/project-access@1.34.3
+
+## 0.6.46
+
+### Patch Changes
+
+- d11943d: fix(deps): update dependency i18next to v25.8.0
+
 ## 0.6.45
 
 ### Patch Changes